--- PJ3T3_Postie/Manager/Utilities.swift
@@ -1,32 +1,0 @@
-//
-//  Utilities.swift
-//  PJ3T3_Postie
-//
-//  Created by Eunsu JEONG on 2/1/24.
-//
-
-import Foundation
-import UIKit
-
-final class Utilities {
-    static let shared = Utilities()
-    private init() { }
-    
-    @MainActor
-    func topViewController(controller: UIViewController? = nil) -> UIViewController? {
-        let controller = controller ?? UIApplication.shared.keyWindow?.rootViewController
-        
-        if let navigationController = controller as? UINavigationController {
-            return topViewController(controller: navigationController.visibleViewController)
-        }
-        if let tabController = controller as? UITabBarController {
-            if let selected = tabController.selectedViewController {
-                return topViewController(controller: selected)
-            }
-        }
-        if let presented = controller?.presentedViewController {
-            return topViewController(controller: presented)
-        }
-        return controller
-    }
-}
